+++ lib/licensed.rb
@@ -5,7 +5,8 @@ require "licensed/dependency_record"
 require "licensed/dependency"
 require "licensed/git"
 require "licensed/sources"
-require "licensed/matchers"
+require "licensed/wildcard_matcher"
+require "licensed/matchers_list"
 require "licensed/configuration"
 require "licensed/reporters"
 require "licensed/commands"
