--- lib/licensed/matchers.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-module Licensed
-  module Matchers
-    require "licensed/matchers/exact"
-    require "licensed/matchers/wildcard"
-    require "licensed/matchers/factory"
-    require "licensed/matchers/list"
-  end
-end
